+++ libmysql/my_alloc.c
@@ -21,7 +21,7 @@
 #include <my_sys.h>
 #include <m_string.h>
 
-void init_alloc_root(MEM_ROOT *mem_root, uint block_size, uint pre_alloc_size)
+void init_alloc_root(MEM_ROOT *mem_root, size_t block_size, size_t pre_alloc_size)
 {
   mem_root->free=mem_root->used=0;
   mem_root->min_malloc=32;
@@ -42,7 +42,7 @@ void init_alloc_root(MEM_ROOT *mem_root, uint block_size, uint pre_alloc_size)
 #endif
 }
 
-gptr alloc_root(MEM_ROOT *mem_root,unsigned int Size)
+gptr alloc_root(MEM_ROOT *mem_root, size_t Size)
 {
 #if defined(HAVE_purify) && defined(EXTRA_DEBUG)
   reg1 USED_MEM *next;
@@ -58,7 +58,7 @@ gptr alloc_root(MEM_ROOT *mem_root,unsigned int Size)
   mem_root->used=next;
   return (gptr) (((char*) next)+ALIGN_SIZE(sizeof(USED_MEM)));
 #else
-  uint get_size,max_left;
+  size_t get_size,max_left;
   gptr point;
   reg1 USED_MEM *next;
   reg2 USED_MEM **prev;
@@ -137,7 +137,7 @@ void free_root(MEM_ROOT *root, myf MyFlags)
 
 char *strdup_root(MEM_ROOT *root,const char *str)
 {
-  uint len= (uint) strlen(str)+1;
+  size_t len= strlen(str)+1;
   char *pos;
   if ((pos=alloc_root(root,len)))
     memcpy(pos,str,len);
@@ -145,7 +145,7 @@ char *strdup_root(MEM_ROOT *root,const char *str)
 }
 
 
-char *memdup_root(MEM_ROOT *root,const char *str,uint len)
+char *memdup_root(MEM_ROOT *root, const char *str, size_t len)
 {
   char *pos;
   if ((pos=alloc_root(root,len)))
