+++ libmysql/strmake.c
@@ -31,7 +31,7 @@
 
 #ifdef BAD_STRING_COMPILER
 
-char *strmake(char *dst,const char *src,uint length)
+char *strmake(char *dst,const char *src, size_t length)
 {
   reg1 char *res;
 
@@ -44,7 +44,7 @@ char *strmake(char *dst,const char *src,uint length)
 #define strmake strmake_overlapp	/* Use orginal for overlapping str */
 #endif
 
-char *strmake(register char *dst, register const char *src, uint length)
+char *strmake(register char *dst, register const char *src, size_t length)
 {
   while (length--)
     if (! (*dst++ = *src++))
